Add evolve monitor and Fix some minor bugs
Evolve monitor is a structure to capture the evolutionary stats such as
best chromosome score of the generation, including its generation number
and chromosome bit string.

In addition to monitoring the stats, a bug was fixed in
evaluate_chromosomes(). The bug did not free the previous score before
setting a new one on top.