Remove genereation stats monitor in memory
Before evolve_monitor was storing all the best chromosomes from each
generation in memory, and then when the GA is finished it sorts all the
data in memory. Well thats kind of stupid. Instead of sorting all 1000s
of elements, evolve_monitor now only keeps the top N-th chromosomes as
defined before execution. This makes it generally simpler for evolve.