Temporary fix on individual memory sizes
At the moment the individual sizes causes problems if they are strings,
this is because in C a string requires a NULL TERMINATOR at the end.
This becomes a problem if one fails to allocate memory for that extral
null terminator, the temporary fix is only for now, a better solution
needs to be thought of to cater for all different sizes.