Omit crossover tests for now
There is a problem in comparing trees that some times the sub trees are
quite similar to each other, and there is currently no good way of
comparing the difference.

Tried comparing addresses however the test makes a copy before and
after, so comparing addresses is out of the question since the addresss
will most certainly be different every time. . .